add of setting command with custom prefix

+++ util/rethinkdb.js
@@ -1,14 +1,14 @@
 const db = module.exports = {};
-
-
+let isLoaded;
 let dbName = "Arys";
 const r = require("rethinkdbdash")({
     host: "192.168.1.30",
     port: "28015",
-    db: "Arys"
+    db: dbName
 });
-
-db.init = async () => {
+const config = require('../config/config');
+db.init = async (Client) => {
+    isLoaded = false;
     let dbs = await r.dbList().run();
     if(!~dbs.indexOf(dbName)) {
         console.info(`Creating database ${dbName}...`);
@@ -17,7 +17,7 @@ db.init = async () => {
 
     let tableList = await r.tableList().run(), tableWait = [];
     let tablesExpected = [
-        "settings", "post", "listenedRoles", "user", "event", "analytic",
+        "setting", "post", "listenedRoles", "user", "event", "analytic",
     ];
     let indexExpected = [
         {table: "analytic", index: "analytic_guild", rows: ["guild"]},
@@ -27,6 +27,7 @@ db.init = async () => {
         {table: "listenedRoles", index: "listenedRoles_guild", rows: ["guild"]},
         {table: "listenedRoles", index: "listenedRoles_guild_role", rows: ["guild", "role"]},
         {table: "listenedRoles", index: "listenedRoles_guild_role_member", rows: ["guild", "role", "member"]},
+        {table: "setting", index: "setting_guild", rows: ["guild"]},
     ];
     let indexes = [];
     for(let table of tablesExpected) {
@@ -40,19 +41,28 @@ db.init = async () => {
         }
     }
     for(let index of indexExpected) {
-        if(!containsObject(index, indexes)) {
+        if(!indexContainsObject(index, indexes)) {
             await r.table(index.table).indexCreate(index.index, index.rows.map(i => r.row(i))).run();
             console.info(`Creating index of "${index.index}" in "${index.table}" table...`);
             await r.table(index.table).indexWait(index.index).run();
             console.info(`Index "${index.index}" in "${index.table}" table is set up`);
         }
     }
+    let guildExpected = Client.guilds.keys(); //array of joinned guild id
+    for(let guild of guildExpected) { //guild is a single guild id
+        let entry = await r.table('setting').getAll([guild], {index: "setting_guild"}).run();
+        if(await entry.length === 0) {
+            await db.createSetting(guild).catch(console.error);
+            console.info(`Guild "${Client.guilds.get(guild).name}" was added in "setting" table`);
+        }
+    }
     await Promise.all(tableWait);
     console.log(`rethinkdb initialized`);
+    isLoaded = true;
     return true;
 };
 
-function containsObject(obj, list) {
+function indexContainsObject(obj, list) {
     for(let i = 0; i < list.length; i++) {
         if (list[i].table === obj.table ) {
             if (list[i].index === obj.index) return true;
@@ -61,7 +71,71 @@ function containsObject(obj, list) {
     return false;
 }
 
+db.createSetting = async (guild) => {
+    if(!isLoaded) return;
+    /*
+     let query = {
+     guild: guild,
+     prefix: String,
+     listenedRoles: Array,
+     //perm: Object,
+     report: Number,
+     responses: Object,
+     lastSave: Date.now(),
+     channel: {
+     log: String,
+     nsfw: Array
+     },
+     mod: {
+     purge: {
+     safe: Boolean,
+     value: Number
+     }
+     }
+     };
+    */
+    let query = {
+        guild: guild,
+        enter: Date.now(),
+        lastSave: Date.now()
+    };
+    return await r.table('setting').insert(query).run();
+};
+
+db.getSetting = async () => {
+    if(!isLoaded) return;
+    let doc =  await r.table('setting').run();
+    let map = new Map();
+    for(let subDoc of doc) {
+        map.set(subDoc.guild, subDoc)
+    }
+    return map;
+};
+
+db.streamSetting = async () => {
+    if(!isLoaded) return;
+    return await r.table('setting').changes().run();
+};
+
+db.setPrefix = async (guild, prefix) => {
+    if(!isLoaded) return;
+    return await r.table('setting').getAll([guild], {index: "setting_guild"}).update({prefix: prefix, lastSave: Date.now()}).run();
+};
+
+db.checkPrefix = async (guild) => {
+    if(!isLoaded) return;
+    let guildObj = await r.table('setting').getAll([guild], {index: "setting_guild"}).run();
+    if(guildObj.length === 0) return false;
+    else if(guildObj.length === 1) return guildObj.first().prefix;
+    else return console.error(`multiple guild found while searching for ${guild}`)
+};
+
+db.deletePrefix = async (guild) => {
+    if(!isLoaded) return;
+    return await r.table('setting').getAll([guild], {index: "setting_guild"}).replace(r.row.without('prefix')).run();
+};
 db.createPost = async (image, message, file, channel, guild) => {
+    if(!isLoaded) return;
     let query = {
         image: image,
         message: message,
@@ -74,19 +148,23 @@ db.createPost = async (image, message, file, channel, guild) => {
 };
 
 db.getPost = async (image, file, guild) => {
+    if(!isLoaded) return;
     return await r.table('post').getAll([guild, file, image], {index: "post_guild_file_image"}).run();
 };
 
 db.reportPost = async (guild, message) => {
+    if(!isLoaded) return;
     let a = await r.table('post').getAll([guild, message], {index: "post_guild_message"}).run();
     return await r.table('post').get(a[0].id).update({report_count: a[0].report_count+1}).run();
 };
 
 db.deletePost = async (guild, message) => {
+    if(!isLoaded) return;
     return await r.table('post').getAll([guild, message], {index: "post_guild_message"}).delete().run();
 };
 
 db.createListenedRole = async (guild, role, member) => {
+    if(!isLoaded) return;
     let query = {
         role: role,
         member: member,
@@ -97,12 +175,14 @@ db.createListenedRole = async (guild, role, member) => {
 };
 
 db.endListenedRole = async (guild, role, member) => {
+    if(!isLoaded) return;
     let doc = await r.table('listenedRoles').getAll([guild, role, member], {index: "listenedRoles_guild_role_member"}).orderBy(r.row("exit")).run();
     return await r.table('listenedRoles').get(doc[0].id).update({exit: Date.now()}).run();
 
 };
 
 db.getListenedRole = async (guild, role, member) => {
+    if(!isLoaded) return;
     if(member === undefined) {
         if(role === undefined) {
             if(guild === undefined) return console.error("please put a guild scope to the query");
@@ -114,6 +194,7 @@ db.getListenedRole = async (guild, role, member) => {
 };
 
 db.createAnalytic = async (guild, channel, item, member, date) => {
+    if(!isLoaded) return;
     let query = {
         item: item,
         user: member,
@@ -125,6 +206,7 @@ db.createAnalytic = async (guild, channel, item, member, date) => {
 };
 
 db.countAnalytic = async (guild) => {
+    if(!isLoaded) return;
     let doc = await r.table('analytic').getAll([guild], {index: "analytic_guild"}).run();
     console.log(doc);
     let nameStack = [...new Set(doc.map(analytic => analytic.item))];
@@ -141,6 +223,7 @@ db.countAnalytic = async (guild) => {
 };
 
 db.countAnalyticByDate = async (guild, min, max) => {
+    if(!isLoaded) return;
     min = min ? new Date(min).getTime() : 0;
     max = max ? new Date(max).getTime() : Date.now();
     console.log("min : " + new Date(min));
@@ -167,24 +250,23 @@ db.countAnalyticByDate = async (guild, min, max) => {
                 let name = nameStack[j];
                 let count = subDoc.filter(function (o) {
                     return (o.item === name);
-            }).length;
-            subStack.push(name + " " + count);
-          }
+                }).length;
+                subStack.push(name + " " + count);
+            }
         }
-      stack[i].value = subStack;
-      stack[i].start = start;
-      stack[i].end = end;
-      }
-      return stack;
+        stack[i].value = subStack;
+        stack[i].start = start;
+        stack[i].end = end;
+    }
+    return stack;
 };
 
 db.fix = async () => {
+    if(!isLoaded) return;
     const mongo = require('./db');
     let doc = await mongo.getAnalytic().catch(console.error);
-    //let doc = await r.table('analytic').run();
     for(let i=0;i<doc.length;i++) {
         if(doc[i].date && doc[i].guild) {
-            //console.log("trigger");
             await db.createAnalytic(doc[i].guild, doc[i].channel, doc[i].item, doc[i].member, parseInt(doc[i].date))
         }
     }
